Improving difficulty progression along levels. Improving explosion performance. Improving asteroids aspect (uniform color).

+++ Xcode/GameManager.h
@@ -26,7 +26,7 @@ namespace asteroids {
 		bool _levelRunning, _cooldownPeriod;
 		double _radarSize;
 		double _cooldownTime, _cooldownLeft;
-		unsigned int _difficulty;
+		double _difficulty;
 
 	public:
 		GameManager(std::string id, MyApp *application);
@@ -52,7 +52,7 @@ namespace asteroids {
 
 #ifndef MY_APP_H
 #define MY_APP_H
-		class StarSphere;
+	class StarSphere;
 	class MyCamera;
 	class GameManager;
 	class ExplosionManager;
